op-replay-clipper-app/requirements.txt
pywebview>=5.0
requests>=2.32
